SwEng02/src/Lab1/Person.java
package Lab1;

public class Person {
	private String name;
	private int age;
	private boolean gender;
	
	public Person() {
		name = "John Doe";
		age = 0;
		gender = true;
		
	}
	
	public Person (String name, int age) {
		this.name = name;
		this.age = age;
	}
	
	public String getName() {
		return name;
	}
	
	public int getAge() {
		return age;
	}
	
	public void setName(String name) {
		this.name = name;
	}
	
	public void setAge(int age) {
		this.age = age;
	}

	public boolean getGender() {
		return gender;
	}

	public void setGender(boolean gender) {
		this.gender = gender;
	}
	
}
